Dont do RIOT yet because of CI issue

+++ include/reactor-uc/serialization.h
@@ -16,5 +16,5 @@ int deserialize_from_protobuf(FederateMessage *message, const unsigned char *buf
 
 lf_ret_t deserialize_payload_default(void *user_struct, const unsigned char *msg_buf, size_t msg_size);
 
-ssize_t serialize_payload_default(const void *user_struct, size_t user_struct_size, unsigned char *msg_buf);
+int serialize_payload_default(const void *user_struct, size_t user_struct_size, unsigned char *msg_buf);
 #endif // REACTOR_UC_SERIALIZATION_H
